fixbug

+++ app/Http/Controllers/Admin/CommentController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 use App\Models\Comment;
@@ -15,7 +16,8 @@ class CommentController extends Controller
 {
     public function index(Request $request)
     {
-    
+
+        // $this->authorize('viewAny', Comment::class);
         $key        = $request->key ?? '';
         $user_id      = $request->user_id ?? '';
         $comment      = $request->comment ?? '';
@@ -51,6 +53,7 @@ class CommentController extends Controller
     }
     public function create()
     {
+        // $this->authorize('create', Comment::class);
         $courses = Course::all();
         $users = User::all();
 
@@ -84,6 +87,7 @@ class CommentController extends Controller
     public function edit($id)
     {
         $comment = Comment::find($id);
+        $this->authorize('update', Comment::class);
 
         // dd($blog);
         return view('Admin.comments.edit', compact('comment'));
@@ -98,12 +102,12 @@ class CommentController extends Controller
 
         try {
             $comment->save();
-            return redirect()->route('comments.index')->with('success', 'Sửa' . ' ' . $request->commentstable_type . ' ' .  'thành công');
+            return redirect()->route('comments.index')->with('success', 'Sửa thành công');
         } catch (\Exception $e) {
             Log::error($e->getMessage());
-            return redirect()->route('comments.index')->with('error', 'Sửa' . ' ' . $request->commentstable_type . ' ' .  'không thành công');
+            return redirect()->route('comments.index')->with('error', 'Sửa không thành công');
         }
-        Session::flash('succes', 'Sửa thành công');
+        Session::flash('success', 'Sửa thành công');
 
         return redirect()->route('comments.index');
 
@@ -111,28 +115,28 @@ class CommentController extends Controller
     public function destroy($id)
     {
         $comment = Comment::find($id);
-
+        $this->authorize('delete', Comment::class);
 
         try {
             $comment->delete();
 
-            return redirect()->route('comments.index')->with('success', 'Xóa' . ' ' . $comment->id . ' ' .  'thành công');
+            return redirect()->route('comments.index')->with('success', 'Xóa thành công');
         } catch (\Exception $e) {
             Log::error($e->getMessage());
-            return redirect()->route('comments.index')->with('error', 'Xóa' . ' ' . $comment->id . ' ' .  'không thành công');
+            return redirect()->route('comments.index')->with('error', 'Xóa không thành công');
         }
     }
     public function force_destroy($id)
     {
 
         $comment = Comment::withTrashed()->find($id);
-
+        $this->authorize('force_destroy',Comment::class);
         try {
             $comment->forceDelete();
-            return redirect()->route('comments.trash')->with('success', 'Xóa' . ' ' . $comment->name . ' ' .  'thành công');
+            return redirect()->route('comments.trash')->with('success', 'Xóa thành công');
         } catch (\Exception $e) {
             Log::error($e->getMessage());
-            return redirect()->route('comments.trash')->with('error', 'Xóa' . ' ' . $comment->name . ' ' .  'không thành công');
+            return redirect()->route('comments.trash')->with('error', 'Xóa không thành công');
         }
     }
 
@@ -140,6 +144,7 @@ class CommentController extends Controller
     {
         date_default_timezone_set("Asia/Ho_Chi_Minh");
         $comment = Comment::findOrFail($id);
+        $this->authorize('force_destroy',Comment::class);
         $comment->deleted_at = date("Y-m-d h:i:s");
         try {
             $comment->save();
@@ -156,6 +161,7 @@ class CommentController extends Controller
     function RestoreDelete($id)
     {
         $comment = Comment::withTrashed()->find($id);
+        $this->authorize('restore',Comment::class);
         $comment->deleted_at = null;
         try {
             $comment->save();
